+++ src/util.ts
@@ -340,6 +340,7 @@ export function getHeightOfRevision(revString: string): number {
 
 import { stringMd5 } from 'pouchdb-md5';
 import { rev as pouchUtilsRev } from 'pouchdb-utils';
+import { BlobBuffer } from './types';
 
 /**
  * Creates a revision string that does NOT include the revision height
@@ -392,3 +393,67 @@ export function isFolderPath(name: string) {
         return false;
     }
 }
+
+
+
+
+export const blobBufferUtil = {
+    /**
+     * depending if we are on node or browser,
+     * we have to use Buffer(node) or Blob(browser)
+     */
+    createBlobBuffer(
+        data: string,
+        type: string
+    ): BlobBuffer {
+        let blobBuffer: any;
+
+        if (isElectronRenderer) {
+            // if we are inside of electron-renderer, always use the node-buffer
+            return Buffer.from(data, {
+                type
+            } as any);
+        }
+
+        try {
+            // for browsers
+            blobBuffer = new Blob([data], {
+                type
+            } as any);
+        } catch (e) {
+            // for node
+            blobBuffer = Buffer.from(data, {
+                type
+            } as any);
+        }
+        return blobBuffer;
+    },
+    toString(blobBuffer: BlobBuffer): Promise<string> {
+        if (blobBuffer instanceof Buffer) {
+            // node
+            return nextTick()
+                .then(() => blobBuffer.toString());
+        }
+        return new Promise(res => {
+            // browsers
+            const reader = new FileReader();
+            reader.addEventListener('loadend', e => {
+                const text = (e.target as any).result;
+                res(text);
+            });
+
+            const blobBufferType = Object.prototype.toString.call(blobBuffer);
+
+            /**
+             * in the electron-renderer we have a typed array insteaf of a blob
+             * so we have to transform it.
+             * @link https://github.com/pubkey/rxdb/issues/1371
+             */
+            if (blobBufferType === '[object Uint8Array]') {
+                blobBuffer = new Blob([blobBuffer]);
+            }
+
+            reader.readAsText(blobBuffer as any);
+        });
+    }
+};